requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4

# Database
sqlalchemy==2.0.31
mysql-connector-python==9.0.0
pymysql==1.1.0
aiomysql==0.2.0
alembic==1.13.2

# Redis for caching
redis==5.0.7

# ChromaDB for RAG
chromadb==1.1.1

# API utilities
httpx==0.27.0
python-multipart==0.0.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS
fastapi-cors==0.0.6

# Logging
loguru==0.7.2

# Date/Time
python-dateutil==2.9.0

# Development
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
black==24.4.2
flake8==7.1.0
mypy==1.11.0
pre-commit==3.7.1